Patches \ Fix popup, add time remaining and a few more fixes
I fixed the popup to correctly be at the center of the screen and the overlay to take the entirety of the screen and. I also added how much time you have before the deadline on tasks and I also fixed the tasksforeach.php html so that the tasks wouldn't be created one inside each other and made it so that the is only put in the html once instead of for every task.